Move custom theme visibility logic to api
